+++ db/schema.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -20,6 +18,7 @@ ActiveRecord::Schema.define(version: 2019_06_01_125350) do
     t.string "website"
     t.string "source"
     t.string "notes"
+    t.integer "visibility"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end