+++ app/controllers/visitors_controller.rb
@@ -2,4 +2,9 @@
 
 # VisitorsController
 class VisitorsController < ApplicationController
+  def index
+    @projects = Project.where(visibility: :everyone).all
+    puts "\n\n@p #{@projects.count}"
+    puts "@everyone #{Project.everyone.first}"
+  end
 end