+++ app/controllers/welcome_controller.rb
@@ -2,5 +2,7 @@
 
 # The signed in users homepage
 class WelcomeController < ApplicationController
-  def index; end
+  def index
+    @projects = Project.all
+  end
 end