+++ app/controllers/projects_controller.rb
@@ -1,0 +1,80 @@
+# frozen_string_literal: true
+
+# ProjectsController
+class ProjectsController < ApplicationController
+  before_action :set_project, only: %i[show edit update destroy]
+  before_action :authenticate_user!
+
+  # GET /projects
+  # GET /projects.json
+  def index
+    @projects = Project.all
+  end
+
+  # GET /projects/1
+  # GET /projects/1.json
+  def show; end
+
+  # GET /projects/new
+  def new
+    @project = Project.new
+  end
+
+  # GET /projects/1/edit
+  def edit; end
+
+  # POST /projects
+  # POST /projects.json
+  def create
+    @project = Project.new(project_params)
+
+    respond_to do |format|
+      if @project.save
+        format.html { success @project, 'created' }
+        format.json { render :show, status: :created, location: @project }
+      else
+        format.html { render :new }
+        format.json { render json: @project.errors, status: :unprocessable_entity }
+      end
+    end
+  end
+
+  # PATCH/PUT /projects/1
+  # PATCH/PUT /projects/1.json
+  def update
+    respond_to do |format|
+      if @project.update(project_params)
+        format.html { success @project, 'updated' }
+        format.json { render :show, status: :ok, location: @project }
+      else
+        format.html { render :edit }
+        format.json { render json: @project.errors, status: :unprocessable_entity }
+      end
+    end
+  end
+
+  # DELETE /projects/1
+  # DELETE /projects/1.json
+  def destroy
+    @project.destroy
+    respond_to do |format|
+      format.html { success projects_url, 'destroyed' }
+      format.json { head :no_content }
+    end
+  end
+
+  private
+
+  def set_project
+    @project = Project.find(params[:id])
+  end
+
+  def project_params
+    params.require(:project)
+          .permit(:name, :outline, :website, :source, :notes, :visibility)
+  end
+
+  def success(destination, message)
+    redirect_to destination, notice: I18n.t("projects.messages.#{message}")
+  end
+end