+++ app/helpers/projects_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# ProjectHelper
+module ProjectsHelper
+end